third_party/blink/web_tests/svg/dynamic-updates/SVGFEComponentTransferElement-svgdom-tableValues-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'tableValues' property of the SVGFEComponentTransferElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS SVGNumberListToString(feRFunc.tableValues.baseVal) is "0 1 1 0 "
PASS SVGNumberListToString(feGFunc.tableValues.baseVal) is "0 1 1 0.5 "
PASS SVGNumberListToString(feBFunc.tableValues.baseVal) is "0 0 1 0.1 "
PASS SVGNumberListToString(feAFunc.tableValues.baseVal) is "0.5 10 1 0.5 "
PASS SVGNumberListToString(feRFunc.tableValues.baseVal) is "0 1 0.9 0 "
PASS SVGNumberListToString(feGFunc.tableValues.baseVal) is "0 1 1 0.6 "
PASS SVGNumberListToString(feBFunc.tableValues.baseVal) is "0 0 1 0.2 "
PASS SVGNumberListToString(feAFunc.tableValues.baseVal) is "0.5 10 1 0.9 "
PASS successfullyParsed is true

TEST COMPLETE
